:hammer: [Refactor] 로그인 구현까지 merge하기 위한 리팩토링
