+++ src/main/java/org/deepdive/apiserver/note/ui/NoteController.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.deepdive.apiserver.common.dto.ResponseDto;
 import org.deepdive.apiserver.note.application.NoteService;
 import org.deepdive.apiserver.note.application.dto.response.GetNoteListResponseDto;
-import org.deepdive.apiserver.security.resolver.Login;
+import org.deepdive.apiserver.security.application.resolver.Login;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
