+++ src/main/java/org/deepdive/apiserver/security/repository/MemberRepository.java
@@ -1,11 +1,13 @@
 package org.deepdive.apiserver.security.repository;
 
 import java.util.Optional;
-import org.deepdive.apiserver.security.entity.Member;
+
+import org.deepdive.apiserver.security.domain.Member;
+import org.deepdive.apiserver.security.repository.entity.MemberEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface MemberRepository extends JpaRepository<Member, Long> {
+public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
 
     Optional<Member> findByEmail(String email);
 }
