+++ src/main/java/org/deepdive/apiserver/security/repository/jpa/JpaMemberRepository.java
@@ -1,0 +1,11 @@
+package org.deepdive.apiserver.security.repository.jpa;
+
+import org.deepdive.apiserver.security.repository.entity.MemberEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface JpaMemberRepository extends JpaRepository<MemberEntity, Long> {
+
+    Optional<MemberEntity> findByEmail(String email);
+}
